+++ hnix-store-core/src/System/Nix/Internal/StorePath.hs
@@ -1,10 +1,9 @@
+{-# OPTIONS_GHC -Wno-orphans #-}
 {-|
 Description : Representation of Nix store paths.
 -}
 {-# language ConstraintKinds #-}
-{-# language RecordWildCards #-}
-{-# language GeneralizedNewtypeDeriving #-}
-{-# language ScopedTypeVariables #-}
+{-# language DeriveAnyClass #-}
 {-# language AllowAmbiguousTypes #-}
 {-# language DataKinds #-}
 
@@ -16,22 +15,31 @@ module System.Nix.Internal.StorePath
   , StorePathHashPart(..)
   , mkStorePathHashPart
   , ContentAddressableAddress(..)
+  , contentAddressableAddressBuilder
+  , contentAddressableAddressParser
+  , digestBuilder
   , NarHashMode(..)
   , -- * Manipulating 'StorePathName'
     makeStorePathName
   , validStorePathName
+    -- * Reason why a path is not valid
+  , InvalidPathError(..)
   , -- * Rendering out 'StorePath's
     storePathToFilePath
   , storePathToRawFilePath
   , storePathToText
   , storePathToNarInfo
+  , storePathHashPartToText
   , -- * Parsing 'StorePath's
     parsePath
   , pathParser
   )
 where
 
+import Data.Default.Class (Default(def))
+import Data.Text.Lazy.Builder (Builder)
 import qualified Relude.Unsafe as Unsafe
+import qualified System.Nix.Hash
 import           System.Nix.Internal.Hash
 import           System.Nix.Internal.Base
 import qualified System.Nix.Internal.Base32    as Nix.Base32
@@ -39,17 +47,23 @@ import qualified System.Nix.Internal.Base32    as Nix.Base32
 import qualified Data.ByteString.Char8         as Bytes.Char8
 import qualified Data.Char                     as Char
 import qualified Data.Text                     as Text
+import qualified Data.Text.Encoding
+import qualified Data.Text.Lazy.Builder
 import           Data.Attoparsec.Text.Lazy      ( Parser
                                                 , (<?>)
                                                 )
+import qualified Data.Attoparsec.ByteString.Char8
 import qualified Data.Attoparsec.Text.Lazy     as Parser.Text.Lazy
 import qualified System.FilePath               as FilePath
 import           Crypto.Hash                    ( SHA256
                                                 , Digest
                                                 , HashAlgorithm
+                                                , hash
                                                 )
 
-import Test.QuickCheck
+import Test.QuickCheck (Arbitrary(arbitrary), listOf, elements)
+import Test.QuickCheck.Arbitrary.Generic (GenericArbitrary(..))
+import Test.QuickCheck.Instances ()
 
 -- | A path in a Nix store.
 --
@@ -68,7 +82,7 @@ data StorePath = StorePath
     -- hello-1.2.3).
     storePathName :: !StorePathName
   }
-  deriving (Eq, Ord, Show)
+  deriving (Eq, Generic, Ord, Show)
 
 instance Hashable StorePath where
   hashWithSalt s StorePath{..} =
@@ -88,7 +102,7 @@ instance Arbitrary StorePath where
 newtype StorePathName = StorePathName
   { -- | Extract the contents of the name.
     unStorePathName :: Text
-  } deriving (Eq, Hashable, Ord, Show)
+  } deriving (Eq, Generic, Hashable, Ord, Show)
 
 instance Arbitrary StorePathName where
   arbitrary = StorePathName . toText <$> ((:) <$> s1 <*> listOf sn)
@@ -102,7 +116,7 @@ newtype StorePathHashPart = StorePathHashPart
   { -- | Extract the contents of the hash.
     unStorePathHashPart :: ByteString
   }
-  deriving (Eq, Hashable, Ord, Show)
+  deriving (Eq, Generic, Hashable, Ord, Show)
 
 instance Arbitrary StorePathHashPart where
   arbitrary = mkStorePathHashPart @SHA256 . Bytes.Char8.pack <$> arbitrary
@@ -112,8 +126,9 @@ mkStorePathHashPart
    . HashAlgorithm hashAlgo
   => ByteString
   -> StorePathHashPart
-mkStorePathHashPart = coerce . mkStorePathHash @hashAlgo
+mkStorePathHashPart = StorePathHashPart . mkStorePathHash @hashAlgo
 
+-- TODO(srk): split into its own module + .Builder/.Parser
 -- | An address for a content-addressable store path, i.e. one whose
 -- store path hash is purely a function of its contents (as opposed to
 -- paths that are derivation outputs, whose hashes are a function of
@@ -132,6 +147,67 @@ data ContentAddressableAddress
     -- addToStore. It is addressed according to some hash algorithm
     -- applied to the nar serialization via some 'NarHashMode'.
     Fixed !NarHashMode !SomeNamedDigest
+  deriving (Eq, Generic, Ord, Show)
+
+-- TODO(srk): extend to all hash types
+instance Arbitrary (Digest SHA256) where
+  arbitrary = hash @ByteString <$> arbitrary
+
+instance Arbitrary SomeNamedDigest where
+  arbitrary = SomeDigest @SHA256 <$> arbitrary
+
+deriving via GenericArbitrary ContentAddressableAddress
+  instance Arbitrary ContentAddressableAddress
+
+-- | Builder for `ContentAddressableAddress`
+contentAddressableAddressBuilder :: ContentAddressableAddress -> Builder
+contentAddressableAddressBuilder (Text digest) =
+  "text:"
+  <> digestBuilder digest
+contentAddressableAddressBuilder (Fixed narHashMode (SomeDigest (digest :: Digest hashAlgo))) =
+  "fixed:"
+  <> (if narHashMode == Recursive then "r:" else mempty)
+--  <> Data.Text.Lazy.Builder.fromText (System.Nix.Hash.algoName @hashAlgo)
+  <> digestBuilder digest
+
+-- | Builder for @Digest@s
+digestBuilder :: forall hashAlgo . (NamedAlgo hashAlgo) => Digest hashAlgo -> Builder
+digestBuilder digest =
+  Data.Text.Lazy.Builder.fromText (System.Nix.Hash.algoName @hashAlgo)
+  <> ":"
+  <> Data.Text.Lazy.Builder.fromText (encodeDigestWith NixBase32 digest)
+
+-- | Parser for content addressable field
+contentAddressableAddressParser :: Data.Attoparsec.ByteString.Char8.Parser ContentAddressableAddress
+contentAddressableAddressParser = caText <|> caFixed
+  where
+  -- | Parser for @text:sha256:<h>@
+  --caText :: Parser ContentAddressableAddress
+  caText = do
+    _      <- "text:sha256:"
+    digest <- decodeDigestWith @SHA256 NixBase32 <$> parseHash
+    either fail pure $ Text <$> digest
+
+  -- | Parser for @fixed:<r?>:<ht>:<h>@
+  --caFixed :: Parser ContentAddressableAddress
+  caFixed = do
+    _           <- "fixed:"
+    narHashMode <- (Recursive <$ "r:") <|> (RegularFile <$ "")
+    digest      <- parseTypedDigest
+    either fail pure $ Fixed narHashMode <$> digest
+
+  --parseTypedDigest :: Parser (Either String SomeNamedDigest)
+  parseTypedDigest = mkNamedDigest <$> parseHashType <*> parseHash
+
+  --parseHashType :: Parser Text
+  parseHashType =
+    Data.Text.Encoding.decodeUtf8
+    <$> ("sha256" <|> "sha512" <|> "sha1" <|> "md5") <* (":" <|> "-")
+
+  --parseHash :: Parser Text
+  parseHash =
+    Data.Text.Encoding.decodeUtf8
+    <$> Data.Attoparsec.ByteString.Char8.takeWhile1 (/= ':')
 
 -- | Schemes for hashing a Nix archive.
 --
@@ -143,19 +219,36 @@ data NarHashMode
   | -- | Hash an arbitrary nar, including a non-executable regular
     -- file if so desired.
     Recursive
+  deriving (Eq, Enum, Generic, Hashable, Ord, Show)
 
-makeStorePathName :: Text -> Either String StorePathName
+deriving via GenericArbitrary NarHashMode
+  instance Arbitrary NarHashMode
+
+-- | Reason why a path is not valid
+data InvalidPathError =
+    EmptyName
+  | PathTooLong
+  | LeadingDot
+  | InvalidCharacter
+  | HashDecodingFailure String
+  | RootDirMismatch
+      { rdMismatchExpected :: StoreDir
+      , rdMismatchGot      :: StoreDir
+      }
+  deriving (Eq, Generic, Hashable, Ord, Show)
+
+makeStorePathName :: Text -> Either InvalidPathError StorePathName
 makeStorePathName n =
   if validStorePathName n
     then pure $ StorePathName n
     else Left $ reasonInvalid n
 
-reasonInvalid :: Text -> String
+reasonInvalid :: Text -> InvalidPathError
 reasonInvalid n
-  | n == ""          = "Empty name"
-  | Text.length n > 211 = "Path too long"
-  | Text.head n == '.'  = "Leading dot"
-  | otherwise        = "Invalid character"
+  | n == ""             = EmptyName
+  | Text.length n > 211 = PathTooLong
+  | Text.head n == '.'  = LeadingDot
+  | otherwise           = InvalidCharacter
 
 validStorePathName :: Text -> Bool
 validStorePathName n =
@@ -183,17 +276,20 @@ type RawFilePath = ByteString
 -- do not know their own store dir by design.
 newtype StoreDir = StoreDir {
     unStoreDir :: RawFilePath
-  } deriving (Eq, Hashable, Ord, Show)
+  } deriving (Eq, Generic, Hashable, Ord, Show)
 
 instance Arbitrary StoreDir where
   arbitrary = StoreDir . ("/" <>) . Bytes.Char8.pack <$> arbitrary
+
+instance Default StoreDir where
+  def = StoreDir "/nix/store"
 
 -- | Render a 'StorePath' as a 'RawFilePath'.
 storePathToRawFilePath :: StoreDir -> StorePath -> RawFilePath
 storePathToRawFilePath storeDir StorePath{..} =
   unStoreDir storeDir <> "/" <> hashPart <> "-" <> name
  where
-  hashPart = encodeUtf8 $ encodeWith NixBase32 $ coerce storePathHash
+  hashPart = encodeUtf8 $ storePathHashPartToText storePathHash
   name     = encodeUtf8 $ unStorePathName storePathName
 
 -- | Render a 'StorePath' as a 'FilePath'.
@@ -210,14 +306,26 @@ storePathToNarInfo :: StorePath -> Bytes.Char8.ByteString
 storePathToNarInfo StorePath{..} =
   encodeUtf8 $ encodeWith NixBase32 (coerce storePathHash) <> ".narinfo"
 
+-- | Render a 'StorePathHashPart' as a 'Text'.
+-- This is used by remote store / database
+-- via queryPathFromHashPart
+storePathHashPartToText :: StorePathHashPart -> Text
+storePathHashPartToText = encodeWith NixBase32 . unStorePathHashPart
+
 -- | Parse `StorePath` from `Bytes.Char8.ByteString`, checking
 -- that store directory matches `expectedRoot`.
-parsePath :: StoreDir -> Bytes.Char8.ByteString -> Either String StorePath
+parsePath
+  :: StoreDir
+  -> Bytes.Char8.ByteString
+  -> Either InvalidPathError StorePath
 parsePath expectedRoot x =
   let
     (rootDir, fname) = FilePath.splitFileName . Bytes.Char8.unpack $ x
     (storeBasedHashPart, namePart) = Text.breakOn "-" $ toText fname
-    storeHash = decodeWith NixBase32 storeBasedHashPart
+    hashPart = bimap
+      HashDecodingFailure
+      StorePathHashPart
+      $ decodeWith NixBase32 storeBasedHashPart
     name = makeStorePathName . Text.drop 1 $ namePart
     --rootDir' = dropTrailingPathSeparator rootDir
     -- cannot use ^^ as it drops multiple slashes /a/b/// -> /a/b
@@ -226,9 +334,12 @@ parsePath expectedRoot x =
     storeDir =
       if expectedRootS == rootDir'
         then pure rootDir'
-        else Left $ "Root store dir mismatch, expected" <> expectedRootS <> "got" <> rootDir'
+        else Left $ RootDirMismatch
+                      { rdMismatchExpected = expectedRoot
+                      , rdMismatchGot = StoreDir $ Bytes.Char8.pack rootDir
+                      }
   in
-    either Left (pure $ StorePath <$> coerce storeHash <*> name) storeDir
+    either Left (pure $ StorePath <$> hashPart <*> name) storeDir
 
 pathParser :: StoreDir -> Parser StorePath
 pathParser expectedRoot = do
@@ -257,8 +368,12 @@ pathParser expectedRoot = do
       <?> "Path name contains invalid character"
 
   let name = makeStorePathName $ Text.cons c0 rest
+      hashPart = bimap
+        HashDecodingFailure
+        StorePathHashPart
+        digest
 
   either
-    fail
+    (fail . show)
     pure
-    (StorePath <$> coerce digest <*> name)
+    (StorePath <$> hashPart <*> name)