+++ hnix-store-core/tests/Derivation.hs
@@ -5,14 +5,18 @@ import           Test.Tasty                     ( TestTree
                                                 , testGroup
                                                 )
 import           Test.Tasty.Golden              ( goldenVsFile )
+import           Test.Tasty.QuickCheck
 
-import           System.Nix.StorePath           ( StoreDir(..) )
+import           Nix.Derivation                 ( Derivation )
+import           System.Nix.StorePath           ( StoreDir(..), StorePath )
 import           System.Nix.Derivation          ( parseDerivation
                                                 , buildDerivation
                                                 )
 
+import Data.Default.Class (Default(def))
 import qualified Data.Attoparsec.Text
 import qualified Data.Text.IO
+import qualified Data.Text.Lazy
 import qualified Data.Text.Lazy.Builder
 
 processDerivation :: FilePath -> FilePath -> IO ()
@@ -24,10 +28,10 @@ processDerivation source dest = do
     (Data.Text.IO.writeFile dest
       . toText
       . Data.Text.Lazy.Builder.toLazyText
-      . buildDerivation (StoreDir "/nix/store")
+      . buildDerivation def
     )
     (Data.Attoparsec.Text.parseOnly
-      (parseDerivation $ StoreDir "/nix/store")
+      (parseDerivation def)
       contents
     )
 
@@ -46,3 +50,17 @@ test_derivation =
     drv = fp <> show n <> ".drv"
     act = fp <> show n <> ".actual"
     fp  = "tests/samples/example"
+
+-- TODO(srk): this won't roundtrip as Arbitrary Text
+-- contains wild stuff like control characters and UTF8 sequences.
+-- Either fix in nix-derivation or use wrapper type
+-- (but we use Nix.Derivation.textParser so we need Text for now)
+xprop_derivationRoundTrip :: StoreDir -> Derivation StorePath Text -> Property
+xprop_derivationRoundTrip = \sd drv ->
+  Data.Attoparsec.Text.parseOnly (parseDerivation sd)
+   ( Data.Text.Lazy.toStrict
+   $ Data.Text.Lazy.Builder.toLazyText
+   $ buildDerivation sd drv
+   )
+  === pure drv
+