+++ hnix-store-remote/src/System/Nix/Store/Remote.hs
@@ -35,7 +35,13 @@ module System.Nix.Store.Remote
   )
 where
 
-import           Prelude                 hiding ( putText )
+import Data.HashSet (HashSet)
+import Data.Map (Map)
+import Data.Text (Text)
+import qualified Control.Monad
+import qualified Data.Attoparsec.ByteString.Char8
+import qualified Data.Text.Encoding
+--
 import qualified Data.ByteString.Lazy          as BSL
 
 import           Nix.Derivation                 ( Derivation )
@@ -50,18 +56,17 @@ import           System.Nix.Hash                ( NamedAlgo(..)
 import           System.Nix.StorePath           ( StorePath
                                                 , StorePathName
                                                 , StorePathHashPart
+                                                , InvalidPathError
                                                 )
-import           System.Nix.StorePathMetadata   ( StorePathMetadata(..)
+import           System.Nix.StorePathMetadata   ( Metadata(..)
                                                 , StorePathTrust(..)
                                                 )
-import           System.Nix.Internal.Base       ( encodeWith )
 
 import qualified Data.Binary.Put
 import qualified Data.Map.Strict
 import qualified Data.Set
 
 import qualified System.Nix.StorePath
-import qualified System.Nix.Store.Remote.Parsers
 
 import           System.Nix.Store.Remote.Binary
 import           System.Nix.Store.Remote.Types
@@ -81,11 +86,11 @@ addToStore
   -> RepairFlag           -- ^ Only used by local store backend
   -> MonadStore StorePath
 addToStore name source recursive repair = do
-  when (unRepairFlag repair)
+  Control.Monad.when (unRepairFlag repair)
     $ error "repairing is not supported when building through the Nix daemon"
 
   runOpArgsIO AddToStore $ \yield -> do
-    yield $ toStrict $ Data.Binary.Put.runPut $ do
+    yield $ BSL.toStrict $ Data.Binary.Put.runPut $ do
       putText $ System.Nix.StorePath.unStorePathName name
       putBool $ not $ System.Nix.Hash.algoName @a == "sha256" && (unRecursive recursive)
       putBool (unRecursive recursive)
@@ -104,7 +109,7 @@ addTextToStore
   -> RepairFlag        -- ^ Repair flag, must be `False` in case of remote backend
   -> MonadStore StorePath
 addTextToStore name text references' repair = do
-  when (unRepairFlag repair)
+  Control.Monad.when (unRepairFlag repair)
     $ error "repairing is not supported when building through the Nix daemon"
 
   storeDir <- getStoreDir
@@ -117,14 +122,14 @@ addTextToStore name text references' repair = do
 addSignatures :: StorePath -> [BSL.ByteString] -> MonadStore ()
 addSignatures p signatures = do
   storeDir <- getStoreDir
-  void $ simpleOpArgs AddSignatures $ do
+  Control.Monad.void $ simpleOpArgs AddSignatures $ do
     putPath storeDir p
     putByteStrings signatures
 
 addIndirectRoot :: StorePath -> MonadStore ()
 addIndirectRoot pn = do
   storeDir <- getStoreDir
-  void $ simpleOpArgs AddIndirectRoot $ putPath storeDir pn
+  Control.Monad.void $ simpleOpArgs AddIndirectRoot $ putPath storeDir pn
 
 -- | Add temporary garbage collector root.
 --
@@ -132,7 +137,7 @@ addIndirectRoot pn = do
 addTempRoot :: StorePath -> MonadStore ()
 addTempRoot pn = do
   storeDir <- getStoreDir
-  void $ simpleOpArgs AddTempRoot $ putPath storeDir pn
+  Control.Monad.void $ simpleOpArgs AddTempRoot $ putPath storeDir pn
 
 -- | Build paths if they are an actual derivations.
 --
@@ -140,7 +145,7 @@ addTempRoot pn = do
 buildPaths :: HashSet StorePath -> BuildMode -> MonadStore ()
 buildPaths ps bm = do
   storeDir <- getStoreDir
-  void $ simpleOpArgs BuildPaths $ do
+  Control.Monad.void $ simpleOpArgs BuildPaths $ do
     putPaths storeDir ps
     putInt $ fromEnum bm
 
@@ -166,7 +171,7 @@ buildDerivation p drv buildMode = do
 ensurePath :: StorePath -> MonadStore ()
 ensurePath pn = do
   storeDir <- getStoreDir
-  void $ simpleOpArgs EnsurePath $ putPath storeDir pn
+  Control.Monad.void $ simpleOpArgs EnsurePath $ putPath storeDir pn
 
 -- | Find garbage collector roots.
 findRoots :: MonadStore (Map BSL.ByteString StorePath)
@@ -177,18 +182,18 @@ findRoots = do
     getSocketIncremental
     $ getMany
     $ (,)
-      <$> (fromStrict <$> getByteStringLen)
+      <$> (BSL.fromStrict <$> getByteStringLen)
       <*> getPath sd
 
   r <- catRights res
   pure $ Data.Map.Strict.fromList r
  where
-  catRights :: [(a, Either String b)] -> MonadStore [(a, b)]
+  catRights :: [(a, Either InvalidPathError b)] -> MonadStore [(a, b)]
   catRights = mapM ex
 
-  ex :: (a, Either [Char] b) -> MonadStore (a, b)
+  ex :: (a, Either InvalidPathError b) -> MonadStore (a, b)
   ex (x , Right y) = pure (x, y)
-  ex (_x, Left e ) = error $ "Unable to decode root: " <> fromString e
+  ex (_x, Left e ) = error $ "Unable to decode root: " <> show e
 
 isValidPathUncached :: StorePath -> MonadStore Bool
 isValidPathUncached p = do
@@ -218,24 +223,24 @@ querySubstitutablePaths ps = do
   runOpArgs QuerySubstitutablePaths $ putPaths storeDir ps
   sockGetPaths
 
-queryPathInfoUncached :: StorePath -> MonadStore StorePathMetadata
+queryPathInfoUncached :: StorePath -> MonadStore (Metadata StorePath)
 queryPathInfoUncached path = do
   storeDir <- getStoreDir
   runOpArgs QueryPathInfo $ do
     putPath storeDir path
 
   valid <- sockGetBool
-  unless valid $ error "Path is not valid"
+  Control.Monad.unless valid $ error "Path is not valid"
 
   deriverPath <- sockGetPathMay
 
-  narHashText <- decodeUtf8 <$> sockGetStr
+  narHashText <- Data.Text.Encoding.decodeUtf8 <$> sockGetStr
   let
     narHash =
       case
         decodeDigestWith @SHA256 NixBase32 narHashText
         of
-        Left  e -> error $ fromString e
+        Left  e -> error e
         Right x -> SomeDigest x
 
   references       <- sockGetPaths
@@ -252,14 +257,16 @@ queryPathInfoUncached path = do
 
       contentAddressableAddress =
         case
-          System.Nix.Store.Remote.Parsers.parseContentAddressableAddress caString
+          Data.Attoparsec.ByteString.Char8.parseOnly
+            System.Nix.StorePath.contentAddressableAddressParser
+            caString
           of
-          Left  e -> error $ fromString e
+          Left  e -> error e
           Right x -> Just x
 
       trust = if ultimate then BuiltLocally else BuiltElsewhere
 
-  pure $ StorePathMetadata{..}
+  pure $ Metadata{..}
 
 queryReferrers :: StorePath -> MonadStore (HashSet StorePath)
 queryReferrers p = do
@@ -288,11 +295,8 @@ queryDerivationOutputNames p = do
 queryPathFromHashPart :: StorePathHashPart -> MonadStore StorePath
 queryPathFromHashPart storePathHash = do
   runOpArgs QueryPathFromHashPart
-    $ putByteStringLen
-    $ encodeUtf8
-    $ encodeWith NixBase32
-    $ System.Nix.StorePath.unStorePathHashPart
-        storePathHash
+    $ putText
+    $ System.Nix.StorePath.storePathHashPartToText storePathHash
   sockGetPath
 
 queryMissing
@@ -316,10 +320,10 @@ queryMissing ps = do
   pure (willBuild, willSubstitute, unknown, downloadSize', narSize')
 
 optimiseStore :: MonadStore ()
-optimiseStore = void $ simpleOp OptimiseStore
+optimiseStore = Control.Monad.void $ simpleOp OptimiseStore
 
 syncWithGC :: MonadStore ()
-syncWithGC = void $ simpleOp SyncWithGC
+syncWithGC = Control.Monad.void $ simpleOp SyncWithGC
 
 -- returns True on errors
 verifyStore :: CheckFlag -> RepairFlag -> MonadStore Bool